+++ core/tracky-mouse.js
@@ -27,13 +27,20 @@ TrackyMouse.loadDependencies = function ({ statsJs = false } = {}) {
 		`${TrackyMouse.dependenciesRoot}/lib/no-eval.js`, // generated with eval-is-evil.html, this instruments clmtrackr.js so I don't need unsafe-eval in the CSP
 		`${TrackyMouse.dependenciesRoot}/lib/clmtrackr.js`,
 		`${TrackyMouse.dependenciesRoot}/lib/face_mesh/face_mesh.js`,
-		`${TrackyMouse.dependenciesRoot}/lib/face-landmarks-detection.min.js`,
 		`${TrackyMouse.dependenciesRoot}/lib/OneEuroFilter.js`,
+	];
+	// face-landmarks-detection.min.js depends on face_mesh.js
+	// avoid sporadic "TypeError: o.Facemesh is not a constructor" by loading face-landmarks-detection after face_mesh.js
+	// TODO: preload in parallel?
+	const moreScriptFiles = [
+		`${TrackyMouse.dependenciesRoot}/lib/face-landmarks-detection.min.js`,
 	];
 	if (statsJs) {
 		scriptFiles.push(`${TrackyMouse.dependenciesRoot}/lib/stats.js`);
 	}
-	return Promise.all(scriptFiles.map(loadScript));
+	return Promise.all(scriptFiles.map(loadScript)).then(() => {
+		return Promise.all(moreScriptFiles.map(loadScript));
+	});
 };
 
 const isSelectorValid = ((dummyElement) =>